Adding annotations in action
