+++ parsers/manifest_parser_test.go
@@ -1595,3 +1595,25 @@ func TestPackageName_Env_Var(t *testing.T) {
         assert.Equal(t, "Apache-2.0", pkg.License, "Get the wrong license.")
     }
 }
+
+func TestComposeActionForAnnotations(t *testing.T) {
+    manifestFile := "../tests/dat/manifest_validate_action_annotations.yaml"
+    mm := NewYAMLParser()
+    manifest, _ := mm.ParseManifest(manifestFile)
+    pkg_name := "packageActionAnnotations"
+    pkg := manifest.Packages[pkg_name]
+    assert.NotNil(t, pkg, "Could not find package with name " + pkg_name)
+    action_name := "helloworld"
+    action := pkg.Actions[action_name]
+    assert.NotNil(t, action, "Could not find action with name " + action_name)
+    actual_annotations := action.Annotations
+    expected_annotations := map[string]interface{} {
+        "action_annotation_1": "this is annotation 1",
+        "action_annotation_2": "this is annotation 2",
+        "action_annotation_3": "this is annotation 3",
+        "action_annotation_4": "this is annotation 44",
+    }
+    assert.Equal(t, len(actual_annotations), len(expected_annotations), "Could not find expected number of annotations specified in manifest file")
+    eq := reflect.DeepEqual(actual_annotations, expected_annotations)
+    assert.True(t, eq, "Expected list of annotations does not match with actual list, expected annotations: %v actual annotations: %v", expected_annotations, actual_annotations)
+}
